ExtractClassProperty - Corrects a bug
When extracting a property if the first thing found by search() is a
constant, then the cursor is position on the ";"
Therefore when we look for the semicolon we end up in a method.
Plus, this search is useless. We are already on the last property
defined at this point so we just have to insert from the current line.